added PR modifications
